Move usage loggers into a new file & add NoopUsageLogger
